Update README.md
While playing this game you will feel the realistic feel of mobile snake game. Which is as same as the mobile snake game only the difference is that the graphics are used in mobile game and we have not used graphics in this project.

The food for the snake is provided at various co-ordinates of the screen to eat. While playing the game, when the snake eats the food automatically the length of the snake is increased by one element. And the score is also increased when the snake eats the food.

In this game, only 3 lives are given to score the points after three life the game will terminate automatically.